Create scheduler.py
This code uses the Tweepy library to authenticate to Twitter's API and send the scheduled tweet. It prompts the user to enter the text of the tweet and the date and time they want the tweet to be sent. It then calculates the time difference between the scheduled time and the current time, and sleeps for that amount of time before sending the tweet. Finally, it prints a message to confirm that the tweet has been successfully scheduled.